+++ tests/test_maidnomadweb/test_organization/test_maid_profile/test_index.py
@@ -5,7 +5,7 @@ import pytest
 class Testメイドさん紹介一覧ページ:
     def test_メイドさんがorder順に表示されること(self, client):
         # arrange
-        from factories import MaidProfileFactory
+        from factories.maidlist import MaidProfileFactory
 
         MaidProfileFactory(
             code="maid1", name="メイド1", order=1, thumbnail_image__filename="maidimg1.jpg"
@@ -42,7 +42,7 @@ class Testメイドさん紹介一覧ページ:
 
     def test_非表示のメイドさんは表示されないこと(self, client):
         # arrange
-        from factories import MaidProfileFactory
+        from factories.maidlist import MaidProfileFactory
 
         MaidProfileFactory(
             code="maid1",
@@ -73,7 +73,7 @@ class Testメイドさん紹介一覧ページ:
 
     def test_サムネイル画像なしの場合メイン画像が表示される(self, client):
         # arrange
-        from factories import MaidProfileFactory
+        from factories.maidlist import MaidProfileFactory
 
         MaidProfileFactory(
             code="maidchan",
@@ -98,7 +98,7 @@ class Testメイドさん紹介一覧ページ:
 
     def test_サムネイル画像もメイン画像もない時はnoimage画像が表示されること(self, client):
         # arrange
-        from factories import MaidProfileFactory
+        from factories.maidlist import MaidProfileFactory
 
         MaidProfileFactory(
             code="maidchan", name="メイドちゃん", thumbnail_image=None, main_image=None
